+++ app/api/common/common.go
@@ -12,7 +12,7 @@ import (
 )
 
 // 수집률로 시각적 성취도 단계 분류
-func PercentCal(percentage uint64) (Code string) {
+func PercentCal(percentage float64) (Code string) {
 
 	switch {
 	case percentage >= 80:
